Add CubeRender class and modyfi Program.cs

+++ Eng_OpenTK/Eng_OpenTK/CubeRender.cs
@@ -22,16 +22,16 @@ namespace Eng_OpenTK
                 partialCount = Math.Pow(partialCount, 3);
 
                 for (int i = 0; i < (int)partialCount; i++)
-                  {
-                      Gl.BindBufferToShaderAttribute(cubesy[i].cube, program, "vertexPosition");
-                      Gl.BindBufferToShaderAttribute(cubesy[i].cubeColor, program, "vertexColor");
-                      Gl.BindBuffer(cubeElements);
+                   {
+                       Gl.BindBufferToShaderAttribute(cubesy[(int)i].cube, program, "vertexPosition");
+                       Gl.BindBufferToShaderAttribute(cubesy[(int)i].cubeColor, program, "vertexColor");
+                       Gl.BindBuffer(cubeElements);
 
-                      Gl.DrawElements(BeginMode.Quads, cubeElements.Count, DrawElementsType.UnsignedInt, IntPtr.Zero);
+                       Gl.DrawElements(BeginMode.Quads, cubeElements.Count, DrawElementsType.UnsignedInt, IntPtr.Zero);
 
 
 
-                  }
+                   }
